[refactor/#23]: ShowTimerResponse 경로 변경 - studyrecord -> timer 로 경로 변경 - 리팩토링
