+++ src/main/java/com/woohaengshi/backend/controller/TimerController.java
@@ -1,6 +1,6 @@
 package com.woohaengshi.backend.controller;
 
-import com.woohaengshi.backend.dto.response.studyrecord.ShowTimerResponse;
+import com.woohaengshi.backend.dto.response.timer.ShowTimerResponse;
 import com.woohaengshi.backend.service.timer.TimerService;
 
 import lombok.RequiredArgsConstructor;
