+++ src/main/java/com/woohaengshi/backend/service/timer/TimerService.java
@@ -1,6 +1,6 @@
 package com.woohaengshi.backend.service.timer;
 
-import com.woohaengshi.backend.dto.response.studyrecord.ShowTimerResponse;
+import com.woohaengshi.backend.dto.response.timer.ShowTimerResponse;
 
 public interface TimerService {
     ShowTimerResponse getTimer(Long memberId);
